Parquet: Add tests for variant metrics.
