+++ api/src/main/java/org/apache/iceberg/variants/SerializedPrimitive.java
@@ -129,6 +129,16 @@ class SerializedPrimitive implements VariantPrimitive<Object>, SerializedValue {
   }
 
   @Override
+  public int hashCode() {
+    return VariantPrimitive.hash(this);
+  }
+
+  @Override
+  public boolean equals(Object other) {
+    return VariantPrimitive.equals(this, other);
+  }
+
+  @Override
   public String toString() {
     return VariantPrimitive.asString(this);
   }
