+++ api/src/main/java/org/apache/iceberg/variants/VariantPrimitive.java
@@ -19,6 +19,7 @@
 package org.apache.iceberg.variants;
 
 import java.nio.ByteBuffer;
+import java.util.Objects;
 import org.apache.iceberg.relocated.com.google.common.io.BaseEncoding;
 import org.apache.iceberg.util.ByteBuffers;
 import org.apache.iceberg.util.DateTimeUtil;
@@ -49,5 +50,22 @@ public interface VariantPrimitive<T> extends VariantValue {
 
   static String asString(VariantPrimitive<?> primitive) {
     return "Variant(type=" + primitive.type() + ", value=" + primitive.valueAsString() + ")";
+  }
+
+  static int hash(VariantPrimitive<?> self) {
+    return Objects.hash(self.type(), self.get());
+  }
+
+  static boolean equals(VariantPrimitive<?> self, Object obj) {
+    if (self == obj) {
+      return true;
+    }
+
+    if (!(obj instanceof VariantPrimitive)) {
+      return false;
+    }
+
+    VariantPrimitive<?> other = (VariantPrimitive<?>) obj;
+    return Objects.equals(self.type(), other.type()) && Objects.equals(self.get(), other.get());
   }
 }
