+++ core/src/main/java/org/apache/iceberg/variants/PrimitiveWrapper.java
@@ -210,6 +210,16 @@ class PrimitiveWrapper<T> implements VariantPrimitive<T> {
   }
 
   @Override
+  public int hashCode() {
+    return VariantPrimitive.hash(this);
+  }
+
+  @Override
+  public boolean equals(Object other) {
+    return VariantPrimitive.equals(this, other);
+  }
+
+  @Override
   public String toString() {
     return VariantPrimitive.asString(this);
   }
