+++ parquet/src/test/java/org/apache/iceberg/parquet/TestVariantWriters.java
@@ -36,26 +36,16 @@ import org.apache.iceberg.inmemory.InMemoryOutputFile;
 import org.apache.iceberg.io.CloseableIterable;
 import org.apache.iceberg.io.FileAppender;
 import org.apache.iceberg.io.OutputFile;
-import org.apache.iceberg.relocated.com.google.common.base.Preconditions;
 import org.apache.iceberg.relocated.com.google.common.collect.ImmutableList;
 import org.apache.iceberg.relocated.com.google.common.collect.ImmutableMap;
 import org.apache.iceberg.relocated.com.google.common.collect.Iterables;
 import org.apache.iceberg.relocated.com.google.common.collect.Lists;
 import org.apache.iceberg.types.Types;
 import org.apache.iceberg.variants.Variant;
-import org.apache.iceberg.variants.VariantArray;
 import org.apache.iceberg.variants.VariantMetadata;
 import org.apache.iceberg.variants.VariantObject;
-import org.apache.iceberg.variants.VariantPrimitive;
 import org.apache.iceberg.variants.VariantTestUtil;
-import org.apache.iceberg.variants.VariantValue;
-import org.apache.iceberg.variants.VariantVisitor;
 import org.apache.iceberg.variants.Variants;
-import org.apache.parquet.schema.GroupType;
-import org.apache.parquet.schema.LogicalTypeAnnotation;
-import org.apache.parquet.schema.PrimitiveType;
-import org.apache.parquet.schema.Type;
-import org.apache.parquet.schema.Types.GroupBuilder;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.FieldSource;
 
@@ -148,7 +138,8 @@ public class TestVariantWriters {
   public void testShreddedValues(Variant variant) throws IOException {
     Record record = RECORD.copy("id", 1, "var", variant);
 
-    Record actual = writeAndRead((id, name) -> toParquetSchema(variant.value()), record);
+    Record actual =
+        writeAndRead((id, name) -> ParquetVariantUtil.toParquetSchema(variant.value()), record);
 
     InternalTestHelpers.assertEquals(SCHEMA.asStruct(), record, actual);
   }
@@ -161,7 +152,9 @@ public class TestVariantWriters {
             .mapToObj(i -> RECORD.copy("id", i, "var", VARIANTS[i]))
             .collect(Collectors.toList());
 
-    List<Record> actual = writeAndRead((id, name) -> toParquetSchema(variant.value()), expected);
+    List<Record> actual =
+        writeAndRead(
+            (id, name) -> ParquetVariantUtil.toParquetSchema(variant.value()), expected);
 
     assertThat(actual.size()).isEqualTo(expected.size());
 
@@ -196,147 +189,6 @@ public class TestVariantWriters {
             .createReaderFunc(fileSchema -> InternalReader.create(SCHEMA, fileSchema))
             .build()) {
       return Lists.newArrayList(reader);
-    }
-  }
-
-  private Type toParquetSchema(VariantValue value) {
-    return VariantVisitor.visit(value, new ParquetSchemaProducer());
-  }
-
-  private static class ParquetSchemaProducer extends VariantVisitor<Type> {
-    @Override
-    public Type object(VariantObject object, List<String> names, List<Type> typedValues) {
-      if (object.numFields() < 1) {
-        // Parquet cannot write  typed_value group with no fields
-        return null;
-      }
-
-      List<GroupType> fields = Lists.newArrayList();
-      int index = 0;
-      for (String name : names) {
-        Type typedValue = typedValues.get(index);
-        fields.add(field(name, typedValue));
-        index += 1;
-      }
-
-      return objectFields(fields);
-    }
-
-    @Override
-    public Type array(VariantArray array, List<Type> elementResults) {
-      throw null;
-    }
-
-    @Override
-    public Type primitive(VariantPrimitive<?> primitive) {
-      switch (primitive.type()) {
-        case NULL:
-          return null;
-        case BOOLEAN_TRUE:
-        case BOOLEAN_FALSE:
-          return shreddedPrimitive(PrimitiveType.PrimitiveTypeName.BOOLEAN);
-        case INT8:
-          return shreddedPrimitive(
-              PrimitiveType.PrimitiveTypeName.INT32, LogicalTypeAnnotation.intType(8));
-        case INT16:
-          return shreddedPrimitive(
-              PrimitiveType.PrimitiveTypeName.INT32, LogicalTypeAnnotation.intType(16));
-        case INT32:
-          return shreddedPrimitive(PrimitiveType.PrimitiveTypeName.INT32);
-        case INT64:
-          return shreddedPrimitive(PrimitiveType.PrimitiveTypeName.INT64);
-        case FLOAT:
-          return shreddedPrimitive(PrimitiveType.PrimitiveTypeName.FLOAT);
-        case DOUBLE:
-          return shreddedPrimitive(PrimitiveType.PrimitiveTypeName.DOUBLE);
-        case DECIMAL4:
-          BigDecimal decimal4 = (BigDecimal) primitive.get();
-          return shreddedPrimitive(
-              PrimitiveType.PrimitiveTypeName.INT32,
-              LogicalTypeAnnotation.decimalType(decimal4.scale(), 9));
-        case DECIMAL8:
-          BigDecimal decimal8 = (BigDecimal) primitive.get();
-          return shreddedPrimitive(
-              PrimitiveType.PrimitiveTypeName.INT64,
-              LogicalTypeAnnotation.decimalType(decimal8.scale(), 18));
-        case DECIMAL16:
-          BigDecimal decimal16 = (BigDecimal) primitive.get();
-          return shreddedPrimitive(
-              PrimitiveType.PrimitiveTypeName.BINARY,
-              LogicalTypeAnnotation.decimalType(decimal16.scale(), 38));
-        case DATE:
-          return shreddedPrimitive(
-              PrimitiveType.PrimitiveTypeName.INT32, LogicalTypeAnnotation.dateType());
-        case TIMESTAMPTZ:
-          return shreddedPrimitive(
-              PrimitiveType.PrimitiveTypeName.INT64,
-              LogicalTypeAnnotation.timestampType(true, LogicalTypeAnnotation.TimeUnit.MICROS));
-        case TIMESTAMPNTZ:
-          return shreddedPrimitive(
-              PrimitiveType.PrimitiveTypeName.INT64,
-              LogicalTypeAnnotation.timestampType(false, LogicalTypeAnnotation.TimeUnit.MICROS));
-        case BINARY:
-          return shreddedPrimitive(PrimitiveType.PrimitiveTypeName.BINARY);
-        case STRING:
-          return shreddedPrimitive(
-              PrimitiveType.PrimitiveTypeName.BINARY, LogicalTypeAnnotation.stringType());
-      }
-
-      throw new UnsupportedOperationException("Unsupported shredding type: " + primitive.type());
-    }
-
-    private static GroupType objectFields(List<GroupType> fields) {
-      GroupBuilder<GroupType> builder =
-          org.apache.parquet.schema.Types.buildGroup(Type.Repetition.OPTIONAL);
-      for (GroupType field : fields) {
-        checkField(field);
-        builder.addField(field);
-      }
-
-      return builder.named("typed_value");
-    }
-
-    private static void checkField(GroupType fieldType) {
-      Preconditions.checkArgument(
-          fieldType.isRepetition(Type.Repetition.REQUIRED),
-          "Invalid field type repetition: %s should be REQUIRED",
-          fieldType.getRepetition());
-    }
-
-    private static GroupType field(String name, Type shreddedType) {
-      GroupBuilder<GroupType> builder =
-          org.apache.parquet.schema.Types.buildGroup(Type.Repetition.REQUIRED)
-              .optional(PrimitiveType.PrimitiveTypeName.BINARY)
-              .named("value");
-
-      if (shreddedType != null) {
-        checkShreddedType(shreddedType);
-        builder.addField(shreddedType);
-      }
-
-      return builder.named(name);
-    }
-
-    private static void checkShreddedType(Type shreddedType) {
-      Preconditions.checkArgument(
-          shreddedType.getName().equals("typed_value"),
-          "Invalid shredded type name: %s should be typed_value",
-          shreddedType.getName());
-      Preconditions.checkArgument(
-          shreddedType.isRepetition(Type.Repetition.OPTIONAL),
-          "Invalid shredded type repetition: %s should be OPTIONAL",
-          shreddedType.getRepetition());
-    }
-
-    private static Type shreddedPrimitive(PrimitiveType.PrimitiveTypeName primitive) {
-      return org.apache.parquet.schema.Types.optional(primitive).named("typed_value");
-    }
-
-    private static Type shreddedPrimitive(
-        PrimitiveType.PrimitiveTypeName primitive, LogicalTypeAnnotation annotation) {
-      return org.apache.parquet.schema.Types.optional(primitive)
-          .as(annotation)
-          .named("typed_value");
     }
   }
 }
